Remove the parameter for EC2 private key
There should be no need to SSH into the EC2 instances, so I've left it commented on the processing-ec2-instances.yml for debugging, in case anyone (or myself) wants to check out inside of the ec2.

I'm keeping the security group with port 22 open tom make future debugging easier, and since it doesn't affects anything.